:recycle: Move OeuvreDb and CalqueDb to models.py
