+++ app/main.py
@@ -1,6 +1,6 @@
 from fastapi import FastAPI, HTTPException, Depends
 from starlette.status import HTTP_404_NOT_FOUND
-from models import Oeuvre,Calque,TypeCalque
+from models import Oeuvre,Calque,TypeCalque,Base,OeuvreDb,CalqueDb
 from datetime import datetime
 from starlette.staticfiles import StaticFiles
 from starlette.responses import Response
@@ -10,8 +10,7 @@ from sqlalchemy import Boolean, Column, Integer, String, create_engine
 from sqlalchemy.ext.declarative import declarative_base, declared_attr
 from sqlalchemy.orm import Session, sessionmaker
 from sqlalchemy import exc
-from sqlalchemy import ForeignKey
-from sqlalchemy.orm import relationship
+
 from sqlalchemy.orm import joinedload
 
 SQLALCHEMY_DATABASE_URI = "sqlite:///./database_ddale.db"
@@ -21,38 +20,6 @@ engine = create_engine(
 )
 SessionLocal = sessionmaker(autocommit=False, autoflush=False, bind=engine)
 
-Base = declarative_base()
-
-# TODO : Ce serait bien de mettre ça dans le fichier models.py.
-# Transcription des classes de models pour les rendre 
-# au même format que la BdD.
-class OeuvreDb(Base):
-    __tablename__ = "oeuvres"
-    id = Column(Integer, primary_key=True, nullable=False)
-    titre = Column(Integer, nullable=False)
-    auteur = Column(String, nullable=False)
-    technique = Column(String, nullable=False)
-    hauteur = Column(Integer, nullable = False)
-    largeur = Column(Integer, nullable = False)
-    annee = Column(Integer, nullable = False)
-    urlCible = Column(String, nullable=False)
-    urlAudio = Column(String)
-    # Le champs 'calques' permet de réaliser la jointure vers les calques.
-    calques = relationship("CalqueDb", back_populates="oeuvre")
-
-
-class CalqueDb(Base):
-    __tablename__ = "calques"
-    id = Column(Integer, primary_key=True, nullable=False)
-    typeCalque = Column(String,nullable=False)
-    description = Column(String, nullable=False)
-    urlCalque = Column(String, nullable=False)
-    urlAudio = Column(String)
-    # Le champs 'oeuvre_id' contient l'id de l'oeuvre.
-    oeuvre_id = Column(Integer, ForeignKey("oeuvres.id"), nullable=False)
-    # Le champs 'oeuvre' permet de réaliser la jointure vers l'oeuvre.
-    oeuvre = relationship("OeuvreDb", back_populates="calques")
-# Faire la jointure dans les deux sens permet de charger le calque à partir de l'oeuvre et vice-versa.
 
 Base.metadata.create_all(bind=engine)
 db_session = SessionLocal()
