+++ app/models.py
@@ -3,6 +3,9 @@ from datetime import datetime
 from typing import List
 from enum import Enum
 from sqlalchemy import Boolean, Column, Integer, String, create_engine
+from sqlalchemy import ForeignKey
+from sqlalchemy.orm import relationship
+from sqlalchemy.ext.declarative import declarative_base
 
 
 class TypeCalque(Enum):
@@ -31,4 +34,37 @@ class Oeuvre(BaseModel):
     urlAudio:UrlStr = Schema ("",description="Url du fichier audio pour le tableau")
 
     calques:List[Calque] = Schema (..., description="Calques contenant des informations sur l'oeuvre")
+
+Base = declarative_base()
+
+# Transcription des classes de models pour les rendre 
+# au même format que la BdD.
+class OeuvreDb(Base):
+    __tablename__ = "oeuvres"
+    id = Column(Integer, primary_key=True, nullable=False)
+    titre = Column(Integer, nullable=False)
+    auteur = Column(String, nullable=False)
+    technique = Column(String, nullable=False)
+    hauteur = Column(Integer, nullable = False)
+    largeur = Column(Integer, nullable = False)
+    annee = Column(Integer, nullable = False)
+    urlCible = Column(String, nullable=False)
+    urlAudio = Column(String)
+    # Le champs 'calques' permet de réaliser la jointure vers les calques.
+    calques = relationship("CalqueDb", back_populates="oeuvre")
+
+
+class CalqueDb(Base):
+    __tablename__ = "calques"
+    id = Column(Integer, primary_key=True, nullable=False)
+    typeCalque = Column(String,nullable=False)
+    description = Column(String, nullable=False)
+    urlCalque = Column(String, nullable=False)
+    urlAudio = Column(String)
+    # Le champs 'oeuvre_id' contient l'id de l'oeuvre.
+    oeuvre_id = Column(Integer, ForeignKey("oeuvres.id"), nullable=False)
+    # Le champs 'oeuvre' permet de réaliser la jointure vers l'oeuvre.
+    oeuvre = relationship("OeuvreDb", back_populates="calques")
+# Faire la jointure dans les deux sens permet de charger le calque à partir de l'oeuvre et vice-versa.
+
